+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 * Unselected images from Batch Edit once User clicks update.
 * Removing key value pair shows entire field being removed in history.
 * Adjust Notification Send Options To Exclude 'Send Now' On Unpublished Content
+* Add YouTube and JW Player video fields
+* Add `aspect_ratio` ui for gallery and document blocks
+* Use `@triniti/schemas` v1.1.6
+* Create `VideoAssetPickerField` and use for video `mezzanine_ref`.
 
 
 ## v0.3.4